Conflicts:
	mem_sim_set.cpp

+++ mem_sim_block.cpp
@@ -2,15 +2,15 @@
 #include "mem_sim_word.hpp"
 #include "mem_sim_exceptions.hpp"
 
-block::block(const unsigned int iword_size, const unsigned int iblock_size) : block_size(iblock_size), valid(0), dirty(0), tag(0), age(0)
+block::block(const unsigned long long int iword_size, const unsigned long long int iblock_size) : block_size(iblock_size), valid(0), dirty(0), tag(0), age(0)
 {
 	words.assign(iblock_size, word(iword_size));
 }
 
-sim_error block::read(std::vector<unsigned int> &data) const
+sim_error block::read(std::vector<unsigned long long int> &data) const
 {
 	sim_error error = Success;
-	for (unsigned int i = 0; i < words.size(); i++)
+	for (unsigned long long int i = 0; i < words.size(); i++)
 	{
 		error = words[i].read(data[i]);
 		if (error)
@@ -20,10 +20,10 @@ sim_error block::read(std::vector<unsigned int> &data) const
 	return Success;
 }
 
-sim_error block::write(const std::vector<unsigned int> &data)
+sim_error block::write(const std::vector<unsigned long long int> &data)
 {
 	sim_error error = Success;
-	for (unsigned int i = 0; i < words.size(); i++)
+	for (unsigned long long int i = 0; i < words.size(); i++)
 	{
 		error = words[i].write(data[i]);
 		if (error)
@@ -43,12 +43,12 @@ bool block::is_dirty() const
 	return dirty;
 }
 
-unsigned int block::get_tag() const
+unsigned long long int block::get_tag() const
 {
 	return tag;
 }
 
-unsigned int block::get_age() const
+unsigned long long int block::get_age() const
 {
 	return age;
 }
