+++ mem_sim_block.hpp
@@ -11,16 +11,16 @@ class block
 {
 public:
 	block(
-		const unsigned int iword_size, // No. of bytes/word
-		const unsigned int iblock_size // No. of words/block
+		const unsigned long long int iword_size, // No. of bytes/word
+		const unsigned long long int iblock_size // No. of words/block
 		);
 
 	sim_error read(
-		std::vector<unsigned int> &data // Variable to store read output in
+		std::vector<unsigned long long int> &data // Variable to store read output in
 		) const;
 
 	sim_error write(
-		const std::vector<unsigned int> &data // Data to write to memory - program will write as many words as there are in data
+		const std::vector<unsigned long long int> &data // Data to write to memory - program will write as many words as there are in data
 		);
 
 	sim_error reset(); // Set all locations to 0
@@ -29,17 +29,17 @@ public:
 
 	bool is_dirty() const;
 
-	unsigned int get_tag() const;
+	unsigned long long int get_tag() const;
 
-	unsigned int get_age() const;
+	unsigned long long int get_age() const;
 
 
 private:
 	bool valid;
 	bool dirty;
-	unsigned int tag;
-	unsigned int age;
-	unsigned int block_size;
+	unsigned long long int tag;
+	unsigned long long int age;
+	unsigned long long int block_size;
 	std::vector<word> words; 
 
 };
